Add mobile detection to the root route to apply to all navigations
