+++ src/server/functions/getIsMobile.ts
@@ -1,0 +1,9 @@
+import { createServerFn } from "@tanstack/react-start";
+import { getRequestHeader } from "@tanstack/react-start/server";
+
+export const getIsMobile = createServerFn().handler(async () => {
+  const userAgent = getRequestHeader("user-agent") || "";
+  const mobileRegex =
+    /Android|webOS|iPhone|iPad|iPod|BlackBerry|IEMobile|Opera Mini/i;
+  return mobileRegex.test(userAgent);
+});
